more code

+++ src/source-code-info.h
@@ -9,8 +9,16 @@ struct SourceCodeInfo {
   std::uint32_t end;
 
   SourceCodeInfo():start(0),end(0){}
-  SourceCodeInfo( std::size_t s , std::size_t e ):start(s),end(e){}
-  SourceCodeInfo( std::uint32_t s , std::uint32_t e ):start(s),end(e){}
+
+  SourceCodeInfo( std::size_t s , std::size_t e ):
+    start(static_cast<std::uint32_t>(s)),
+    end  (static_cast<std::uint32_t>(e))
+  {}
+
+  SourceCodeInfo( std::uint32_t s , std::uint32_t e ):
+    start(s),
+    end  (e)
+  {}
 };
 
 static_assert( sizeof(SourceCodeInfo) == 8 );
